Splited up frontpy.py
